s3/vfs_tsmsm: fix debug output for large files (cherry picked from commit d5d08d3b50592f8950ab68c654b2116e704d0f30)

+++ source3/modules/vfs_tsmsm.c
@@ -163,10 +163,10 @@ static bool tsmsm_is_offline(struct vfs_handle_struct *handle,
 	   then assume it is not offline (it may not be 100%, as it could be sparse) */
 	if (512 * (off_t)stbuf->st_ex_blocks >=
 	    stbuf->st_ex_size * tsmd->online_ratio) {
-		DEBUG(10,("%s not offline: st_blocks=%ld st_size=%ld "
+		DEBUG(10,("%s not offline: st_blocks=%llu st_size=%llu "
 			  "online_ratio=%.2f\n", path,
-			  (long)stbuf->st_ex_blocks,
-			  (long)stbuf->st_ex_size, tsmd->online_ratio));
+			  (unsigned long long)stbuf->st_ex_blocks,
+			  (unsigned long long)stbuf->st_ex_size, tsmd->online_ratio));
 		return false;
 	}
 
@@ -263,9 +263,9 @@ static bool tsmsm_aio_force(struct vfs_handle_struct *handle, struct files_struc
 	   if the file might be offline
 	*/
 	if(SMB_VFS_FSTAT(fsp, &sbuf) == 0) {
-		DEBUG(10,("tsmsm_aio_force st_blocks=%ld st_size=%ld "
-			  "online_ratio=%.2f\n", (long)sbuf.st_ex_blocks,
-			  (long)sbuf.st_ex_size, tsmd->online_ratio));
+		DEBUG(10,("tsmsm_aio_force st_blocks=%llu st_size=%llu "
+			  "online_ratio=%.2f\n", (unsigned long long)sbuf.st_ex_blocks,
+			  (unsigned long long)sbuf.st_ex_size, tsmd->online_ratio));
 		return !(512 * (off_t)sbuf.st_ex_blocks >=
 			 sbuf.st_ex_size * tsmd->online_ratio);
 	}
